move particles funcs to separate module
